Update test (#109)
